requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
Pillow>=10.0.0
pydrive2>=1.16.0
streamlit-gsheets>=0.1.5
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0